requirements_new.txt
Flask==2.3.3
Flask-MongoEngine==1.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
WeasyPrint==60.2
xhtml2pdf==0.2.8
python-docx==1.1.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
mongoengine==0.27.0
